refactor: move app config to root of @scow/config
