+++ apps/auth/src/config/clusters.ts
@@ -1,3 +1,3 @@
-import { getClusterConfigs } from "@scow/config/build/appConfig/cluster";
+import { getClusterConfigs } from "@scow/config/build/cluster";
 
 export const clusters = getClusterConfigs();
