+++ apps/auth/src/config/ui.ts
@@ -1,3 +1,3 @@
-import { getUiConfig } from "@scow/config/build/appConfig/ui";
+import { getUiConfig } from "@scow/config/build/ui";
 
 export const uiConfig = getUiConfig();
