+++ apps/mis-server/src/config/clusters.ts
@@ -1,4 +1,4 @@
-import { getClusterConfigs } from "@scow/config/build/appConfig/cluster";
+import { getClusterConfigs } from "@scow/config/build/cluster";
 
 export const clusters = getClusterConfigs();
