+++ apps/mis-web/package.json
@@ -25,6 +25,7 @@
     "@ddadaal/tsgrpc-client": "0.17.1",
     "@grpc/grpc-js": "1.7.3",
     "@scow/config": "workspace:*",
+    "@scow/lib-config": "workspace:*",
     "@scow/lib-auth": "workspace:*",
     "@scow/lib-decimal": "workspace:*",
     "@sinclair/typebox": "0.25.2",
