+++ apps/portal-server/src/config/env.ts
@@ -1,4 +1,4 @@
-import { envConfig, host, num, port, str } from "@scow/config";
+import { envConfig, host, num, port, str } from "@scow/lib-config";
 import { getKeyPair } from "@scow/lib-ssh";
 import { homedir } from "os";
 import { join } from "path";
