+++ apps/portal-server/src/config/portal.ts
@@ -1,4 +1,4 @@
-import { getPortalConfig } from "@scow/config/build/appConfig/portal";
+import { getPortalConfig } from "@scow/config/build/portal";
 
 export const portalConfig = getPortalConfig();
