+++ dockerfiles/Dockerfile.mis-web
@@ -9,6 +9,7 @@ WORKDIR /app
 COPY package.json pnpm-lock.yaml pnpm-workspace.yaml .npmrc ./
 
 COPY libs/auth/package.json ./libs/auth/
+COPY libs/libconfig/package.json ./libs/libconfig/
 COPY libs/config/package.json ./libs/config/
 COPY libs/decimal/package.json ./libs/decimal/
 COPY apps/mis-web/package.json apps/mis-web/
@@ -19,6 +20,7 @@ COPY tsconfig.json .eslintrc.json ./
 
 COPY protos ./protos
 
+COPY libs/libconfig ./libs/libconfig
 COPY libs/config ./libs/config
 COPY libs/auth ./libs/auth
 COPY libs/decimal ./libs/decimal
@@ -41,10 +43,16 @@ COPY --from=builder /app/libs/auth/package.json .
 COPY --from=builder /app/libs/auth/node_modules/ ./node_modules
 COPY --from=builder /app/libs/auth/build/ ./build
 
+WORKDIR /app/libs/libconfig
+COPY --from=builder /app/libs/libconfig/package.json .
+COPY --from=builder /app/libs/libconfig/node_modules/ ./node_modules
+COPY --from=builder /app/libs/libconfig/build/ ./build
+
 WORKDIR /app/libs/config
 COPY --from=builder /app/libs/config/package.json .
 COPY --from=builder /app/libs/config/node_modules/ ./node_modules
 COPY --from=builder /app/libs/config/build/ ./build
+
 
 WORKDIR /app/apps/mis-web
 COPY --from=builder /app/apps/mis-web/package.json ./package.json
