+++ dockerfiles/Dockerfile.portal-server
@@ -8,6 +8,7 @@ WORKDIR /app
 
 COPY package.json pnpm-lock.yaml pnpm-workspace.yaml .npmrc ./
 
+COPY libs/libconfig/package.json ./libs/libconfig/
 COPY libs/config/package.json ./libs/config/
 COPY libs/decimal/package.json ./libs/decimal/
 COPY libs/ssh/package.json ./libs/ssh/
@@ -19,6 +20,7 @@ COPY tsconfig.json .eslintrc.json ./
 
 COPY protos ./protos
 
+COPY libs/libconfig ./libs/libconfig
 COPY libs/config ./libs/config
 COPY libs/auth ./libs/auth
 COPY libs/decimal ./libs/decimal
@@ -39,15 +41,20 @@ COPY --from=builder /app/libs/decimal/package.json .
 COPY --from=builder /app/libs/decimal/node_modules/ ./node_modules
 COPY --from=builder /app/libs/decimal/build/ ./build
 
-WORKDIR /app/libs/config
-COPY --from=builder /app/libs/config/package.json .
-COPY --from=builder /app/libs/config/node_modules/ ./node_modules
-COPY --from=builder /app/libs/config/build/ ./build
+WORKDIR /app/libs/libconfig
+COPY --from=builder /app/libs/libconfig/package.json .
+COPY --from=builder /app/libs/libconfig/node_modules/ ./node_modules
+COPY --from=builder /app/libs/libconfig/build/ ./build
 
 WORKDIR /app/libs/ssh
 COPY --from=builder /app/libs/ssh/package.json .
 COPY --from=builder /app/libs/ssh/node_modules/ ./node_modules
 COPY --from=builder /app/libs/ssh/build/ ./build
+
+WORKDIR /app/libs/config
+COPY --from=builder /app/libs/config/package.json .
+COPY --from=builder /app/libs/config/node_modules/ ./node_modules
+COPY --from=builder /app/libs/config/build/ ./build
 
 WORKDIR /app/apps/portal-server
 COPY --from=builder /app/apps/portal-server/package.json .
