+++ dockerfiles/Dockerfile.portal-web
@@ -9,6 +9,7 @@ WORKDIR /app
 COPY package.json pnpm-lock.yaml pnpm-workspace.yaml .npmrc ./
 
 COPY libs/auth/package.json ./libs/auth/
+COPY libs/libconfig/package.json ./libs/libconfig/
 COPY libs/config/package.json ./libs/config/
 COPY libs/decimal/package.json ./libs/decimal/
 COPY libs/ssh/package.json ./libs/ssh/
@@ -24,6 +25,7 @@ COPY tsconfig.json .eslintrc.json ./
 
 COPY protos ./protos
 
+COPY libs/libconfig ./libs/libconfig
 COPY libs/config ./libs/config
 COPY libs/auth ./libs/auth
 COPY libs/decimal ./libs/decimal
@@ -46,6 +48,12 @@ WORKDIR /app/libs/auth
 COPY --from=builder /app/libs/auth/package.json .
 COPY --from=builder /app/libs/auth/node_modules/ ./node_modules
 COPY --from=builder /app/libs/auth/build/ ./build
+
+
+WORKDIR /app/libs/libconfig
+COPY --from=builder /app/libs/libconfig/package.json .
+COPY --from=builder /app/libs/libconfig/node_modules/ ./node_modules
+COPY --from=builder /app/libs/libconfig/build/ ./build
 
 WORKDIR /app/libs/config
 COPY --from=builder /app/libs/config/package.json .
