+++ libs/config/package.json
@@ -2,7 +2,6 @@
   "name": "@scow/config",
   "version": "1.0.0",
   "description": "",
-  "main": "build/index.js",
   "scripts": {
     "build": "rimraf build && tsc -p tsconfig.build.json && tsc-alias -p tsconfig.build.json"
   },
