+++ libs/config/src/constants.ts
@@ -1,5 +1,4 @@
-export const CONFIG_BASE_PATH = "/etc/scow";
+export const PRODUCTION_CONFIG_BASE_PATH = "/etc/scow";
 
-/** @internal */
-export const defaultBasePathConfig = process.env.NODE_ENV === "production" ? CONFIG_BASE_PATH : "config";
+export const DEFAULT_CONFIG_BASE_PATH = process.env.NODE_ENV === "production" ? PRODUCTION_CONFIG_BASE_PATH : "config";
